Refactor: 따로 Extension 파일들을 모으지말고 사용하는 곳에 속성을 추가하거나 extension을 이동시킴
