--- PhotosApp/PhotosApp/Extensions/Notification.swift
@@ -1,16 +1,0 @@
-//
-//  Notification.swift
-//  PhotosApp
-//
-//  Created by kimdo2297 on 2020/03/19.
-//  Copyright © 2020 jinie. All rights reserved.
-//
-
-import Foundation
-
-extension Notification.Name {
-    
-    static let notificationPhotoLibraryDidChange = Notification.Name("photoLibraryDidChange")
-    static let notificationDoodleImageDidChange = Notification.Name("doodleImageDidChange")
-    
-}
